fixing bug with annual interval

+++ backend/pkg/preventative_care/model.go
@@ -1,16 +1,16 @@
 package preventative_care
 
 type RecommendedPreventativeCareItem struct {
-	Id             int    `json:"id"`
-	SearchTerm     string `json:"searchTerm"`
-	Title          string `json:"title"`
-	Description    string `json:"description"`
-	Gender         string `json:"gender"`
-	AgeRangeMin    int    `json:"ageRangeMin"`
-	AgeRangeMax    int    `json:"ageRangeMax"`
-	SmokingHistory bool   `json:"smokingHistory"`
-	Recurring      bool   `json:"isRecurring"`
-	AnnualInterval int    `json:"annualInterval"`
+	Id             int     `json:"id"`
+	SearchTerm     string  `json:"searchTerm"`
+	Title          string  `json:"title"`
+	Description    string  `json:"description"`
+	Gender         string  `json:"gender"`
+	AgeRangeMin    int     `json:"ageRangeMin"`
+	AgeRangeMax    int     `json:"ageRangeMax"`
+	SmokingHistory bool    `json:"smokingHistory"`
+	Recurring      bool    `json:"isRecurring"`
+	AnnualInterval float64 `json:"annualInterval"`
 }
 
 type RecommendedPreventativeCareItemRequest struct {
